+++ src/index.js
@@ -1,5 +1,5 @@
-export { default as IntentDetector } from "./intentDetector";
-export { default as RandomRunner } from "./runner/randomRunner";
-export { default as ConfigurationBuilder } from "./configurationBuilder"
+export { default as IntentDetector } from "./intent_detector";
+export { default as RandomRunner } from "./runner/random_runner";
+export { default as ConfigurationBuilder } from "./configuration_builder"
 export { default as Configuration } from "./configuration"
-export { default as DialogueManger } from "./dialogueManager"
+export { default as DialogueManger } from "./dialogue_manager"
