+++ src/matcher/state_condition.js
